[update]company 하위 접근이 안되서 따로 뺌
