+++ lib/routes/wholesale.ts
@@ -87,52 +87,6 @@ router.put("/:id", async (req: Request, res: Response) => {
 });
 
 
-//Company
-router.get("/company", async (req, res) => {
-    console.log("/company get")
-    companyModel.findAll((err: Error, company: Company[]) => {
-        if (err) {
-            return res.status(500).json({ "errorMessage": err.message });
-        }
-
-        res.status(200).json({ "data": company });
-    });
-});
-
-router.post("/company", async (req: Request, res: Response) => {
-    console.log("/company post")
-    const newCompany: BasicCompany = req.body;
-    companyModel.create(newCompany, (err: Error, companyId: number) => {
-        if (err) {
-            return res.status(500).json({ "message": err.message });
-        }
-        res.status(200).json({ "companyId": companyId });
-    });
-});
-
-router.get("/company/:id", async (req: Request, res: Response) => {
-    console.log("/company get id")
-    const companyId: number = Number(req.params.id);
-    companyModel.findOne(companyId, (err: Error, company: Company) => {
-        if (err) {
-            return res.status(500).json({ "message": err.message });
-        }
-        res.status(200).json({ "data": company });
-    })
-});
-
-router.put("/company/:id", async (req: Request, res: Response) => {
-    console.log("/company put")
-    const company: Company = req.body;
-    companyModel.update(company, (err: Error) => {
-        if (err) {
-            return res.status(500).json({ "message": err.message });
-        }
-        res.status(200).send();
-    })
-});
-
-
 
 // router.get('/', (req, res, next) => {
 //     var bPayment = req.query.bPayment;
